backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
nltk
yfinance
praw
python-dotenv
textblob